python_package_requirements.txt
Flask==2.2.3
mysql-connector-python==8.3.0
web3==5.31.3
werkzeug==2.2.3